Move Component type definition to account package
