+++ pkg/util/util.go
@@ -10,26 +10,6 @@ import (
 	"github.com/jaegertracing/jaeger-operator/pkg/apis/jaegertracing/v1"
 )
 
-// Component represents type of jaeger component
-type Component string
-
-const (
-	// CollectorComponent represents the value for the Component type for Jaeger Collector
-	CollectorComponent Component = "collector"
-
-	// QueryComponent represents the value for the Component type for Jaeger Query
-	QueryComponent Component = "query"
-
-	// IngesterComponent represents the value for the Component type for Jaeger Ingester
-	IngesterComponent Component = "ingester"
-
-	// AllInOneComponent represents the value for the Component type for Jaeger All-In-One
-	AllInOneComponent Component = "all-in-one"
-
-	// AgentComponent epresents the value for the Component type for Jaeger Agent
-	AgentComponent Component = "agent"
-)
-
 // removeDuplicatedVolumes returns a unique list of Volumes based on Volume names. Only the first item is kept.
 func removeDuplicatedVolumes(volumes []corev1.Volume) []corev1.Volume {
 	var results []corev1.Volume
